Major bug fix
Fixed numerous issues with the code, including an incorrect win-check function namely for the horizontal win case, but also due to a missing curly brace that caused a much larger and obfusctaed for loop than necessary (not to mention that it caused one of the inner for loops to use the same iterator variable as the outer for loop, which was a major problem). Other problems that were fixed included a missing code block that would allow the computer to go first when playing if the user did not choose 'X' or 'x' as their character of choice, a lack of flushing the input buffer that caused some issues with reading input, and a couple of unnecessary commented-out lines that made the code less readable.